added authentication and version control

+++ app/services/entity.py
@@ -1,21 +1,29 @@
 from app.core.database import mongodb
 from bson import ObjectId
-from datetime import datetime
+from datetime import datetime, timedelta
 from pymongo.collection import ReturnDocument
 from pymongo.results import UpdateResult
 from app.core.database import mongodb
-
+from app.schemas.entity import EntityUpdate
+from fastapi import Request, HTTPException, status, Depends
+from jose import jwt, JWTError
+import copy
 
 def get_entity_collection():
     if mongodb.db is None:
         raise RuntimeError("Database not initialized. Did you forget to call connect_to_mongo?")
     return mongodb.db["entities"]
 
+def get_entity_history_collection():
+    if mongodb.db is None:
+        raise RuntimeError("Database not initialized. Did you forget to call connect_to_mongo?")
+    return mongodb.db["entity_history"]
 
-def create_entity(data: dict):
+async def create_entity(data: dict):
     collection = get_entity_collection()
     data["created_at"] = datetime.now()
     result = collection.insert_one(data)
+    await save_history(data, "create")
     return str(result.inserted_id)
 
 def list_entities():
@@ -30,7 +38,6 @@ def list_entities():
 def get_entity_by_id(entity_id: str):
     collection = get_entity_collection()
     try:
-        # Validate that entity_id is a valid ObjectId
         object_id = ObjectId(entity_id)
         entity = collection.find_one({"_id": object_id})
         if entity:
@@ -38,7 +45,6 @@ def get_entity_by_id(entity_id: str):
             del entity["_id"]
         return entity
     except Exception:
-        # If entity_id is not a valid ObjectId, return None
         return None
 
 def get_entity_by_name(entity_name: str):
@@ -49,28 +55,66 @@ def get_entity_by_name(entity_name: str):
         del entity["_id"]
     return entity
 
-def update_entity(entity_id: str, entity_data: dict) -> bool:
-    collection = mongodb.db["entities"]
+async def update_entity(entity_id: str, update_data: EntityUpdate):
+    collection = get_entity_collection()
+    existing = collection.find_one({"_id": ObjectId(entity_id)})
+    if not existing:
+        raise HTTPException(status_code=404, detail="Entity not found")
 
-    # Remove fields with None (so they are not updated)
-    update_data = {k: v for k, v in entity_data.model_dump(exclude_unset=True).items() if v is not None}
+    new_version = existing.get("version", 1) + 1
+    update_doc = update_data.model_dump(exclude_unset=True)
+    update_doc["version"] = new_version
+    updated_entity = {**existing, **update_doc}
+    await save_history(updated_entity, operation="update")
 
-    if not update_data:
-        return False  # Nothing to update
-
-    result: UpdateResult = collection.update_one(
+    collection.update_one(
         {"_id": ObjectId(entity_id)},
-        {"$set": update_data}
+        {"$set": update_doc}
     )
-    return result.modified_count == 1
 
-def delete_entity(entity_id: str):
+    updated = collection.find_one({"_id": ObjectId(entity_id)})
+    return updated
+
+async def save_history(entity: dict, operation: str):
+    entity_history_collection = get_entity_history_collection()
+    history_doc = {
+        "entity_id": entity["_id"],
+        "version": entity.get("version", 1),
+        "data": copy.deepcopy(entity),
+        "operation": operation,
+        "timestamp": datetime.utcnow()
+    }
+
+    entity_history_collection.insert_one(history_doc)
+
+
+async def delete_entity(entity_id: str):
     collection = get_entity_collection()
     try:
-        # Validate that entity_id is a valid ObjectId
         object_id = ObjectId(entity_id)
+        entity = collection.find_one({"_id": object_id})
+        if not entity:
+            return False
         result = collection.delete_one({"_id": object_id})
-        return result.deleted_count == 1
-    except Exception:
-        # If entity_id is not a valid ObjectId, return False
+        if result.deleted_count == 1:
+            await save_history(entity, "delete")
+            return True
         return False
+
+    except Exception:
+        return False
+
+def serialize_doc(doc):
+    for key, value in doc.items():
+        if isinstance(value, ObjectId):
+            doc[key] = str(value)
+        elif isinstance(value, dict):
+            doc[key] = serialize_doc(value)
+    return doc
+
+def get_entity_history_by_id(entity_id: str):
+    history_collection = get_entity_history_collection()
+    object_id = ObjectId(entity_id)
+    cursor = history_collection.find({"entity_id": object_id}).sort("version", 1)
+    history = [serialize_doc(doc) for doc in cursor]
+    return history
